+++ cat-consumer/src/main/java/com/dianping/cat/consumer/problem/AbstractProblemHandler.java
@@ -1,4 +1,4 @@
-package org.unidal.cat.plugin.problem;
+package com.dianping.cat.consumer.problem;
 
 import com.dianping.cat.consumer.problem.model.entity.Duration;
 import com.dianping.cat.consumer.problem.model.entity.Entity;
