--- cat-consumer/src/main/java/org/unidal/cat/plugin/problem/ProblemConstants.java
@@ -1,5 +1,0 @@
-package org.unidal.cat.plugin.problem;
-
-public interface ProblemConstants {
-   String NAME = "problem";
-}
