+++ cat-consumer/src/main/resources/META-INF/plexus/components.xml
@@ -1,144 +1,6 @@
 <plexus>
 	<components>
 		<component>
-			<role>org.unidal.cat.spi.analysis.pipeline.Pipeline</role>
-			<role-hint>problem</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.ProblemPipeline</implementation>
-			<instantiation-strategy>per-lookup</instantiation-strategy>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.spi.analysis.MessageRoutingStrategy</role>
-					<role-hint>domain-hash</role-hint>
-				</requirement>
-				<requirement>
-					<role>org.unidal.cat.spi.ReportConfiguration</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.ReportManager</role>
-			<role-hint>problem</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.ProblemReportManager</implementation>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.spi.report.provider.ReportProvider</role>
-				</requirement>
-				<requirement>
-					<role>org.unidal.cat.spi.report.storage.ReportStorage</role>
-					<field-name>m_storage</field-name>
-				</requirement>
-				<requirement>
-					<role>org.unidal.cat.spi.report.storage.ReportStorage</role>
-					<role-hint>file</role-hint>
-					<field-name>m_fileStorage</field-name>
-				</requirement>
-				<requirement>
-					<role>org.unidal.cat.spi.report.internals.ReportDelegateManager</role>
-				</requirement>
-				<requirement>
-					<role>org.unidal.cat.spi.report.ReportFilterManager</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.report.ReportAggregator</role>
-			<role-hint>problem</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.ProblemReportAggregator</implementation>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.report.ReportDelegate</role>
-			<role-hint>problem</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.ProblemReportDelegate</implementation>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.spi.report.ReportAggregator</role>
-					<role-hint>problem</role-hint>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.analysis.MessageAnalyzer</role>
-			<role-hint>problem</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.ProblemReportAnalyzer</implementation>
-			<instantiation-strategy>per-lookup</instantiation-strategy>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.plugin.problem.ProblemHandler</role>
-					<role-hints>
-						<role-hint>default</role-hint>
-						<role-hint>long-execution</role-hint>
-					</role-hints>
-					<field-name>m_handlers</field-name>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.plugin.problem.ProblemHandler</role>
-			<implementation>org.unidal.cat.plugin.problem.DefaultProblemHandler</implementation>
-			<requirements>
-				<requirement>
-					<role>com.dianping.cat.config.server.ServerConfigManager</role>
-				</requirement>
-				<requirement>
-					<role>java.util.Set</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.plugin.problem.ProblemHandler</role>
-			<role-hint>long-execution</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.LongExecutionProblemHandler</implementation>
-			<requirements>
-				<requirement>
-					<role>com.dianping.cat.config.server.ServerConfigManager</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.plugin.problem.filter.ProblemReportHelper</role>
-			<implementation>org.unidal.cat.plugin.problem.filter.ProblemReportHelper</implementation>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.report.ReportFilter</role>
-			<role-hint>problem:homepage</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.filter.ProblemHomePageFilter</implementation>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.plugin.problem.filter.ProblemReportHelper</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.report.ReportFilter</role>
-			<role-hint>problem:graph</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.filter.ProblemGraphFilter</implementation>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.plugin.problem.filter.ProblemReportHelper</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.report.ReportFilter</role>
-			<role-hint>problem:thread</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.filter.ProblemThreadFilter</implementation>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.plugin.problem.filter.ProblemReportHelper</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
-			<role>org.unidal.cat.spi.report.ReportFilter</role>
-			<role-hint>problem:detail</role-hint>
-			<implementation>org.unidal.cat.plugin.problem.filter.ProblemDetailFilter</implementation>
-			<requirements>
-				<requirement>
-					<role>org.unidal.cat.plugin.problem.filter.ProblemReportHelper</role>
-				</requirement>
-			</requirements>
-		</component>
-		<component>
 			<role>com.dianping.cat.analysis.MessageAnalyzer</role>
 			<role-hint>heartbeat</role-hint>
 			<implementation>com.dianping.cat.consumer.heartbeat.HeartbeatAnalyzer</implementation>
