+++ cat-home/src/main/java/com/dianping/cat/report/page/problem/transform/ProblemStatistics.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
-import com.dianping.cat.consumer.problem.ProblemType;
+import org.unidal.cat.plugin.problem.ProblemType;
 import com.dianping.cat.consumer.problem.model.entity.Duration;
 import com.dianping.cat.consumer.problem.model.entity.Entity;
 import com.dianping.cat.consumer.problem.model.entity.Machine;
