+++ cat-plugin-problem/pom.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <parent>
+        <groupId>com.dianping.cat</groupId>
+        <artifactId>parent</artifactId>
+        <version>2.0-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>cat-plugin-problem</artifactId>
+    <name>CAT :: Plugin :: Problem</name>
+    <packaging>jar</packaging>
+    <dependencies>
+        <dependency>
+            <groupId>com.dianping.cat</groupId>
+            <artifactId>cat-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.unidal.framework</groupId>
+            <artifactId>test-framework</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.unidal.maven.plugins</groupId>
+                <artifactId>codegen-maven-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>generate dal model files</id>
+                        <phase>generate-sources</phase>
+                        <goals>
+                            <goal>dal-model</goal>
+                        </goals>
+                        <configuration>
+                            <manifest>
+                                ${basedir}/src/main/resources/META-INF/dal/model/problem-report-manifest.xml,
+                            </manifest>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>org.unidal.maven.plugins</groupId>
+                <artifactId>plexus-maven-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>generate plexus component descriptor</id>
+                        <phase>process-classes</phase>
+                        <goals>
+                            <goal>plexus</goal>
+                        </goals>
+                        <configuration>
+                            <className>org.unidal.cat.plugin.problem.build.ComponentsConfigurator</className>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+    <properties>
+        <project.build.sourceEncoding>utf-8</project.build.sourceEncoding>
+    </properties>
+</project>
